+++ integrations/google_ai/src/haystack_integrations/components/generators/google_ai/chat/gemini.py
@@ -242,10 +242,15 @@ class GoogleAIGeminiChatGenerator:
             p = Part()
             p.text = message.text
             return p
-        elif message.is_from(ChatRole.FUNCTION):
+        elif "FUNCTION" in ChatRole._member_names_ and message.is_from(ChatRole.FUNCTION):
             p = Part()
             p.function_response.name = message.name
             p.function_response.response = message.text
+            return p
+        elif message.is_from(ChatRole.TOOL):
+            p = Part()
+            p.function_response.name = message.tool_call_result.origin.tool_name
+            p.function_response.response = message.tool_call_result.result
             return p
         elif message.is_from(ChatRole.USER):
             return self._convert_part(message.text)
@@ -260,16 +265,23 @@ class GoogleAIGeminiChatGenerator:
         elif message.is_from(ChatRole.SYSTEM) or message.is_from(ChatRole.ASSISTANT):
             part = Part()
             part.text = message.text
-        elif message.is_from(ChatRole.FUNCTION):
+        elif "FUNCTION" in ChatRole._member_names_ and message.is_from(ChatRole.FUNCTION):
             part = Part()
             part.function_response.name = message.name
             part.function_response.response = message.text
         elif message.is_from(ChatRole.USER):
             part = self._convert_part(message.text)
+        elif message.is_from(ChatRole.TOOL):
+            part = Part()
+            part.function_response.name = message.tool_call_result.origin.tool_name
+            part.function_response.response = message.tool_call_result.result
         else:
             msg = f"Unsupported message role {message.role}"
             raise ValueError(msg)
-        role = "user" if message.is_from(ChatRole.USER) or message.is_from(ChatRole.FUNCTION) else "model"
+
+        role = "user"
+        if message.is_from(ChatRole.ASSISTANT) or message.is_from(ChatRole.SYSTEM):
+            role = "model"
         return Content(parts=[part], role=role)
 
     @component.output_types(replies=List[ChatMessage])
@@ -335,13 +347,16 @@ class GoogleAIGeminiChatGenerator:
 
             for part in candidate.content.parts:
                 if part.text != "":
-                    replies.append(ChatMessage.from_assistant(content=part.text, meta=candidate_metadata))
+                    replies.append(ChatMessage.from_assistant(part.text, meta=candidate_metadata))
                 elif part.function_call:
                     candidate_metadata["function_call"] = part.function_call
                     new_message = ChatMessage.from_assistant(
-                        content=json.dumps(dict(part.function_call.args)), meta=candidate_metadata
+                        json.dumps(dict(part.function_call.args)), meta=candidate_metadata
                     )
-                    new_message.name = part.function_call.name
+                    try:
+                        new_message.name = part.function_call.name
+                    except AttributeError:
+                        new_message._name = part.function_call.name
                     replies.append(new_message)
         return replies
 
@@ -364,12 +379,15 @@ class GoogleAIGeminiChatGenerator:
                 for part in candidate["content"]["parts"]:
                     if "text" in part and part["text"] != "":
                         content = part["text"]
-                        replies.append(ChatMessage.from_assistant(content=content, meta=metadata))
+                        replies.append(ChatMessage.from_assistant(content, meta=metadata))
                     elif "function_call" in part and len(part["function_call"]) > 0:
                         metadata["function_call"] = part["function_call"]
                         content = json.dumps(dict(part["function_call"]["args"]))
-                        new_message = ChatMessage.from_assistant(content=content, meta=metadata)
-                        new_message.name = part["function_call"]["name"]
+                        new_message = ChatMessage.from_assistant(content, meta=metadata)
+                        try:
+                            new_message.name = part["function_call"]["name"]
+                        except AttributeError:
+                            new_message._name = part["function_call"]["name"]
                         replies.append(new_message)
 
                     streaming_callback(StreamingChunk(content=content, meta=metadata))