+++ integrations/google_ai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [integrations/google_ai-v4.1.0] - 2025-01-16
+
+### 🧹 Chores
+
+- Google-ai - gently handle the removal of function role (#1297)
+
+
+## [integrations/google_ai-v4.0.1] - 2024-12-19
+
+### 🐛 Bug Fixes
+
+- Make GoogleAI Chat Generator compatible with new `ChatMessage`; small fixes to Cohere tests (#1253)
+
+
 ## [integrations/google_ai-v4.0.0] - 2024-12-10
 
 ### 🐛 Bug Fixes