+++ integrations/langfuse/CHANGELOG.md
@@ -1,6 +1,30 @@
 # Changelog
 
-## [unreleased]
+## [integrations/langfuse-v0.7.0] - 2025-01-21
+
+### 🚀 Features
+
+- LangfuseConnector - add httpx.Client init param (#1308)
+
+### 🐛 Bug Fixes
+
+- End langfuse generation spans properly (#1301)
+
+
+## [integrations/langfuse-v0.6.4] - 2025-01-17
+
+### 🚀 Features
+
+- Add LangfuseConnector secure key management and serialization  (#1287)
+
+
+## [integrations/langfuse-v0.6.3] - 2025-01-15
+
+### 🌀 Miscellaneous
+
+- Chore: Langfuse - pin `haystack-ai>=2.9.0` and simplify message conversion (#1292)
+
+## [integrations/langfuse-v0.6.2] - 2025-01-02
 
 ### 🚀 Features
 
@@ -13,6 +37,7 @@
 ### 🌀 Miscellaneous
 
 - Chore: Fix tracing_context_var lint errors (#1220)
+- Fix messages conversion to OpenAI format (#1272)
 
 ## [integrations/langfuse-v0.6.0] - 2024-11-18