+++ integrations/ollama/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [integrations/ollama-v2.2.0] - 2025-01-16
+
+### 🚀 Features
+
+- Ollama - add support for tools (#1294)
+
+
+## [integrations/ollama-v2.1.2] - 2024-12-18
+
+### 🐛 Bug Fixes
+
+- Make Ollama Chat Generator compatible with new ChatMessage (#1256)
+
+
 ## [integrations/ollama-v2.1.1] - 2024-12-10
 
 ### 🌀 Miscellaneous